cigar alignment inside parser_out script
