+++ tests/unit_tests/test_hmmer_parser_out.py
@@ -5,12 +5,9 @@ pytest -v
 """
 
 import json
-import os
-import sys
 import pytest
 
 
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../../interproscan/scripts/hmmer')))
 from interproscan.scripts.hmmer.parser_out import parse
